refactor(users): move favorite clinic methods to user service
