+++ src/main/java/com/medspace/application/usecase/tenantFavoriteClinic/AssignTenantFavoriteClinicToTenantUseCase.java
@@ -1,6 +1,6 @@
 package com.medspace.application.usecase.tenantFavoriteClinic;
 
-import com.medspace.application.service.TenantFavoriteClinicService;
+import com.medspace.application.service.UserService;
 import com.medspace.domain.model.TenantFavoriteClinic;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
@@ -8,9 +8,9 @@ import jakarta.inject.Inject;
 @ApplicationScoped
 public class AssignTenantFavoriteClinicToTenantUseCase {
     @Inject
-    TenantFavoriteClinicService tenantFavoriteClinicService;
+    UserService userService;
 
     public TenantFavoriteClinic execute(Long favoriteClinicId, Long tenantId) {
-        return tenantFavoriteClinicService.assignToTenant(favoriteClinicId, tenantId);
+        return userService.assignFavoriteClinicToTenant(favoriteClinicId, tenantId);
     }
 }
