+++ src/main/java/com/medspace/application/usecase/tenantFavoriteClinic/GetFavoriteClinicByIdUseCase.java
@@ -1,6 +1,6 @@
 package com.medspace.application.usecase.tenantFavoriteClinic;
 
-import com.medspace.application.service.TenantFavoriteClinicService;
+import com.medspace.application.service.UserService;
 import com.medspace.domain.model.TenantFavoriteClinic;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
@@ -8,9 +8,9 @@ import jakarta.inject.Inject;
 @ApplicationScoped
 public class GetFavoriteClinicByIdUseCase {
     @Inject
-    TenantFavoriteClinicService tenantFavoriteClinicService;
+    UserService userService;
 
     public TenantFavoriteClinic execute(Long id) {
-        return tenantFavoriteClinicService.getFavoriteClinicById(id);
+        return userService.getTenantFavoriteClinicById(id);
     }
 }
