+++ src/main/java/com/medspace/application/usecase/tenantFavoriteClinic/GetFavoriteClinicsByClinicIdUseCase.java
@@ -1,6 +1,6 @@
 package com.medspace.application.usecase.tenantFavoriteClinic;
 
-import com.medspace.application.service.TenantFavoriteClinicService;
+import com.medspace.application.service.UserService;
 import com.medspace.domain.model.TenantFavoriteClinic;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
@@ -10,9 +10,9 @@ import java.util.List;
 @ApplicationScoped
 public class GetFavoriteClinicsByClinicIdUseCase {
     @Inject
-    TenantFavoriteClinicService tenantFavoriteClinicService;
+    UserService userService;
 
     public List<TenantFavoriteClinic> execute(Long clinicId) {
-        return tenantFavoriteClinicService.getFavoriteClinicsByClinicId(clinicId);
+        return userService.getFavoriteClinicsByClinicId(clinicId);
     }
 }
