+++ src/main/java/com/medspace/application/usecase/tenantFavoriteClinic/GetFavoriteClinicsByTenantIdUseCase.java
@@ -1,6 +1,6 @@
 package com.medspace.application.usecase.tenantFavoriteClinic;
 
-import com.medspace.application.service.TenantFavoriteClinicService;
+import com.medspace.application.service.UserService;
 import com.medspace.domain.model.TenantFavoriteClinic;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
@@ -10,9 +10,9 @@ import java.util.List;
 @ApplicationScoped
 public class GetFavoriteClinicsByTenantIdUseCase {
     @Inject
-    TenantFavoriteClinicService tenantFavoriteClinicService;
+    UserService userService;
 
     public List<TenantFavoriteClinic> execute(Long tenantId) {
-        return tenantFavoriteClinicService.getFavoriteClinicsByTenantId(tenantId);
+        return userService.getFavoriteClinicsByTenantId(tenantId);
     }
 }
