+++ src/main/java/com/medspace/application/usecase/tenantFavoriteClinic/IsFavoriteClinicUseCase.java
@@ -1,15 +1,15 @@
 package com.medspace.application.usecase.tenantFavoriteClinic;
 
-import com.medspace.application.service.TenantFavoriteClinicService;
+import com.medspace.application.service.UserService;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
 
 @ApplicationScoped
 public class IsFavoriteClinicUseCase {
     @Inject
-    TenantFavoriteClinicService tenantFavoriteClinicService;
+    UserService userService;
 
     public boolean execute(Long tenantId, Long clinicId) {
-        return tenantFavoriteClinicService.isFavoriteClinic(tenantId, clinicId);
+        return userService.isTenantFavoriteClinic(tenantId, clinicId);
     }
 }
